pyproject.toml
[tool.poetry]
name = "openbook"
version = "0.1.0"
description = "OpenBook: Interactive Online Textbooks"
authors = ["Dennis Schulmeister-Zimolong <dennis@wpvs.de>"]
homepage = "https://openbook.studio"
repository = "https://github.com/DennisSchulmeister/openbook"
license = "AGPL-3.0"
readme = "README.md"

# Use poetry for dependency management but not for packaging, since at least
# for now we are not publishing a python packages on PyPI. Maybe we want to do
# in future, in which case we need to reconsider some of the project structure.
package-mode = false

[tool.poetry.dependencies]
channels = {extras = ["daphne"], version = "^4.3.2"}
channels-redis = "^4.3.0"
daphne = "^4.2.1"
Django = "^6.0.3"
django-allauth = {extras = ["headless-spec", "saml", "socialaccount"], version = "^65.15.0"}
django-cleanup = "^9.0.0"
django-crispy-forms = "^2.6"
django-dbbackup = "^5.2.0"
django-extensions = "^4.1"
django-filter = "^25.2"
django-import-export = {extras = ["all"], version = "^4.4.0"}
django-unfold = "^0.85.0"
djangoql = "^0.19.1"
djangorestframework = "^3.16.1"
drf-flex-fields2 = "^2.0.2"
drf-spectacular = {extras = ["sidecar"], version = "^0.29.0"}
markdown = "^3.10.2"
pillow = "^12.1.1"
pygments = "^2.19.2"
python = "^3.12"
pyyaml = "^6.0.3"
semver = "^3.0.4"
twisted = {extras = ["http2", "tls"], version = "^25.5.0"}
django-colorfield = "^0.14.0"

[tool.poetry.group.dev.dependencies]
tox = "^4.50.0"
pip-audit = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
